Change UndeployPowerAction behaviour
UndeployPowerAction is documented here:
http://pubs.vmware.com/vcd-55/index.jsp?topic=%2Fcom.vmware.vcloud.api.reference.doc_55%2Fdoc%2Ftypes%2FUndeployVAppParamsType.html

There are a number of values that can be passed to it, including Shutdown
and PowerOff. PowerOff and Shutdown differ in that the former is the
equivalent of pulling the power out at the wall on a physical machine, and
Shutdown performs a graceful shutdown of the vApp which then performs a
graceful shutdown of the VM. The default action for the UndeployPowerAction
call is PowerOff.

Until now, we've been relying on the default value of PowerOff rather than
Shutdown, which isn't the nicest way to turn machines off. Fixes this to
make clean boots a thing.